un-commenting tests in datamanager/tests/service_test.cc + reflect datamanager changes to the tests

+++ src/maidsafe/vault/tests/tests_utils.h
@@ -43,8 +43,8 @@ namespace vault {
 
 namespace test {
 
-static const size_t kTestChunkSize = 1024 * 1024;
-static const size_t kAverageChunksStored = 1000;
+static const uint64_t kTestChunkSize = 1024 * 1024;
+static const uint64_t kAverageChunksStored = 1000;
 
 routing::NodeInfo MakeNodeInfo(const passport::Pmid& pmid);
